--- style.css
@@ -1,239 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-  background: rgba(18, 18, 18, 0.9); 
-  background-image: url('images/bakground.webp'); 
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  color: white;
-  overflow-x: hidden;
-  height: 100vh;
-
-  backdrop-filter: blur(8px);
-  -webkit-backdrop-filter: blur(8px);
-}
-
-
-
-.interface-wrapper {
-  display: flex;
-  height: 100vh;
-  width: 100%;
-  position: relative;
-}
-
-.center-box {
-  margin: auto;
-  text-align: center;
-  max-width: 700px;
-}
-
-h1 {
-  margin-bottom: 50px;
-}
-
-.case-opener {
-  position: relative;
-  width: 600px;
-  height: 120px;
-  overflow: hidden;
-  border: 2px solid #444;
-  background: #222;
-  margin: 0 auto 15px auto;
-}
-
-.case-image {
-  width:: 256px;
-  height: 256px;
-}
-
-.carousel {
-  display: flex;
-  height: 100%;
-  transition: transform 5s cubic-bezier(0.1, 0.8, 0.3, 1);
-  will-change: transform;
-}
-
-.carousel img {
-  height: 100px;
-  width: 100px;
-  margin-right: 10px;
-  object-fit: contain;
-}
-
-.pointer {
-  position: absolute;
-  top: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 20px;
-  height: 100%;
-  background-color: #5bc0ff;
-  clip-path: polygon(50% 0%, 100% 50%, 50% 100%, 0% 50%);
-  z-index: 10;
-  pointer-events: none;
-}
-
-
-button {
-  margin-top: 5px; 
-  padding: 12px 28px;
-  font-size: 16px;
-  cursor: pointer;
-
-  background-color: #5bc0ff;
-  color: white;
-  border: none;
-  border-radius: 12px;
-  transition: background-color 0.3s ease, transform 0.2s ease;
-  box-shadow: 0 4px 8px rgba(91, 192, 255, 0.5);
-}
-
-button:hover {
-  background-color: #47aaff;
-  transform: scale(1.05);
-}
-
-.result {
-  margin-top: 15px;
-  font-size: 1.4rem;
-  font-weight: bold;
-}
-
-.history-panel {
-  position: fixed;
-  top: 0;
-  right: 0;
-  width: 250px;
-  height: 100vh;
-  background: rgba(18, 18, 18, 0.5);
-  border-left: 2px solid rgba(255, 255, 255, 0.1);
-  padding: 20px;
-  box-sizing: border-box;
-  overflow-y: auto;
-  color: white;
-
-  backdrop-filter: blur(8px);
-  -webkit-backdrop-filter: blur(8px);
-}
-
-
-.history-panel h2 {
-  font-size: 1.2rem;
-  margin-bottom: 10px;
-}
-
-.history-list {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.history-item {
-  display: flex;
-  align-items: center;
-  background: #2a2a2a;
-  border: 1px solid #444;
-  border-radius: 6px;
-  padding: 5px;
-  gap: 10px;
-}
-
-.history-item img {
-  width: 40px;
-  height: 40px;
-  object-fit: contain;
-}
-
-.image-wrapper {
-  position: relative;
-  display: inline-block;
-}
-
-.rarity-bar {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  height: 4px;
-  width: 100%;
-  border-radius: 0 0 4px 4px;
-}
-
-.blue-bar {
-  background-color: #007bff;
-}
-
-.purple-bar {
-  background-color: #a200ff;
-}
-
-.pink-bar {
-  background-color: #ff5ec2;
-}
-
-.yellow-bar {
-  background-color: #ffd700;
-}
-
-.chances-panel {
-  position: fixed;
-  top: 20px;
-  left: 20px;
-  width: 230px;
-  background: rgba(0, 0, 0, 0.6);
-  padding: 15px 20px;
-  border-radius: 16px;
-  color: white;
-  font-family: sans-serif;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
-  backdrop-filter: blur(6px);
-  -webkit-backdrop-filter: blur(6px);
-}
-
-.chances-panel h3 {
-  margin-top: 0;
-  font-size: 18px;
-  margin-bottom: 10px;
-  text-align: center;
-}
-
-.chances-panel ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.chances-panel li {
-  display: flex;
-  align-items: center;
-  margin-bottom: 8px;
-  font-size: 15px;
-}
-
-.dot {
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  display: inline-block;
-  margin-right: 10px;
-}
-
-.dot.common {
-  background-color: #4299e1; /* blue */
-}
-
-.dot.rare {
-  background-color: #9f7aea; /* purple */
-}
-
-.dot.epic {
-  background-color: #ed64a6; /* pink */
-}
-
-.dot.legendary {
-  background-color: #f6e05e; /* yellow */
-}
-
-
